moved addStoreImage component to new file StoreImageFrom, renamed AddStore Component and file to StoreFrom
